feat(render): 后端帧边界切换逻辑（D3 §12.5）Vello 失败→Skia fallback
- RenderBackend trait 新增 as_any_mut() 下转型方法
- VelloBackend 覆写 as_any_mut() 以暴露后端特定操作
- AppHandler render_ctx 从 VelloBackend 改为 Box<dyn RenderBackend>
- 新增 backend_fallback_pending 标志位实现帧边界切换
- Vello 渲染失败标记 fallback，下一帧通过 RenderBackendFactory 切换
- resize/set_atlas_data 等 Vello 特定操作通过下转型处理
- skia-backend 加入 rgui 默认 features 确保 fallback 可用
- 修复 skia.rs clippy unnecessary_cast 警告

+++ rgui-render/src/skia.rs
@@ -332,7 +332,7 @@ impl RenderBackend for SkiaBackend {
     ) {
         // CPU 纹理更新：在存储的纹理数据中替换子区域。
         if let Some(tex) = self.textures.get_mut(&id) {
-            let stride = tex.width as u32 * 4;
+            let stride = tex.width * 4;
             for row in 0..height {
                 let src_start = (row * width * 4) as usize;
                 let src_end = src_start + (width * 4) as usize;
